+++ VesselWebCenter/VesselWebCenter.Services/ViewModels/VoyageDataViewModel.cs
@@ -13,7 +13,7 @@
         public string DestPortLong { get; set; } = null!;
         public string Country { get; set; } = null!;
         public string UNLocode { get; set; } = null!;
-        public int ExpectedSpeed { get; set; }
+        public double ExpectedSpeed { get; set; }
         public double CalculatedDistance { get; set; }
         public double CalculatedTime => this.CalculatedDistance / (this.ExpectedSpeed*1.0);
         public DateTime DepartureTime => DateTime.UtcNow;
